user_service/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
click==8.1.6
dnspython==2.4.1
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pika==1.3.2
pika-stubs==0.1.3
pluggy==1.2.0
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.4.1
pytest==7.4.0
pytest-tempdir==2019.10.12
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
